+++ R/mu2.R
@@ -27,7 +27,114 @@ mu2env$expit <- rxode2::expit
     expr
   }
 }
+
+#' Get mu3 covariate
+#'
+#' This function evaluates a given expression within a specific
+#' environment and data context.
+#'
+#' @param data The data frame that will be optimized, will contain the
+#'   covariates built into the expression to built a new data column
+#'   for traditional mu referencing
+#'
+#' @param .datEnv The environment that contains the symengine
+#'   environment used to gather expressions and help evaluate them in
+#'   the context of the data and model.
+#'
+#' @param .tmp A string representing the expression to be evaluated to
+#'   see if it will produce a new dataset column without errors.
+#'
+#' @return The evaluated result of the expression
+#'
+#' @noRd
+#' @keywords internal
+#' @author Matthew L. Fidler
 .uiGetMu3 <- function(data, .datEnv, .tmp) {
+  .tmp <- eval(str2lang(paste0("rxode2::rxToSE(", .tmp, ", NULL)")))
+  .tmp <- str2lang(paste0("with(.datEnv$symengine, ", .tmp, ")"))
+  .tmp <- eval(.tmp)
+  .tmp <- as.character(.tmp)
+  .tmp <- str2lang(paste0("rxode2::rxFromSE(", .tmp, ")"))
+  .tmp <- eval(.tmp)
+  .tmp <- str2lang(paste0("with(.datEnv, with(data,",  .tmp, "))"))
+  eval(.tmp)
+}
+
+#' Process and transform a call to change string assigment to integer
+#' assigments according to `rxode2`'s rules
+#'
+#' This function recursively processes a call object `x` and
+#' transforms it based on the provided string assignments in
+#' `strAssign`.  It specifically looks for equality (`==`) and
+#' inequality (`!=`) comparisons and replaces parts of the call with
+#' numeric indices from `strAssign`.
+#'
+#' This is used to transform mu4 covariates to mu covariates.
+#'
+#' @param x A call object to be processed.
+#'
+#' @param strAssign A named list of string assignments used for transformation.
+#'
+#' @return A transformed call object with parts replaced by numeric indices.
+#'
+#' @noRd
+.uiGetMu4f <- function(x, strAssign) {
+  if (is.call(x)) {
+    if (identical(x[[1]], quote(`==`)) ||
+          identical(x[[1]], quote(`!=`))) {
+      .d1 <- deparse1(x[[2]])
+      .w <- which(.d1 == names(strAssign))
+      if (length(.w) == 1L)  {
+        .d2 <- x[[3]]
+        .w <- which(.d2 == strAssign[[.d1]])
+        if (length(.w) == 1) {
+          x[[3]] <- as.numeric(.w)
+          return(x)
+        }
+      } else {
+        .d2 <- deparse1(x[[3]])
+        .w <- which(.d2 == names(strAssign))
+        if (length(.w) == 1L) {
+          .d1 <- x[[2]]
+          .w <- which(.d1 == strAssign[[.d2]])
+          if (length(.w) == 1L) {
+            x[[2]] <- as.numeric(.w)
+            return(x)
+          }
+        }
+      }
+    }
+    return(as.call(c(x[[1]],lapply(x[-1], .uiGetMu4f, strAssign=strAssign))))
+  }
+  x
+}
+
+#' @title Internal Function to Process mu4 expressions with symengine
+#'   and string replacement
+#' @description This function processes the input data using the
+#'   SymEngine environment and the `rxode2` package
+#'
+#' @param data The data to be processed; if it has the right
+#'   information it will create a new column to be added as a
+#'   traditional mu-expression in `nlmixr2`
+#'
+#' @param .datEnv The environment containing the symengine which is
+#'   the rxode2 model translated for use with symengine. Also in this
+#'   environment is the ui which contains the rxode2 ui model
+#'
+#' @param .tmp A temporary expression to be evaluated within the
+#'   SymEngine context.
+#'
+#' @return The result of evaluating the processed expression within
+#'   the data environment.
+#' @noRd
+#' @keywords internal
+#' @author Matthew L. Fidler
+.uiGetMu4 <- function(data, .datEnv, .tmp) {
+  .sa <- rxode2::rxModelVars(.datEnv$ui)$strAssign
+  .lang <- str2lang(.tmp)
+  .lang <-.uiGetMu4f(.lang, .sa)
+  .tmp <- deparse1(.lang)
   .tmp <- eval(str2lang(paste0("rxode2::rxToSE(", .tmp, ", NULL)")))
   .tmp <- str2lang(paste0("with(.datEnv$symengine, ", .tmp, ")"))
   .tmp <- eval(.tmp)
@@ -56,13 +163,16 @@ mu2env$expit <- rxode2::expit
   if (use.utf()) {
     .mu2 <- "\u03BC\u2082"
     .mu3 <- "\u03BC\u2083"
+    .mu4 <- "\u03BC\u2084"
   } else {
     .mu2 <- "mu2"
     .mu3 <- "mu3"
+    .mu4 <- "mu4"
   }
-
   lapply(seq_along(ui$mu2RefCovariateReplaceDataFrame$covariate),
          function(i) {
+           .bad <- paste0("not ",.mu2,", ", .mu3, " or ", .mu4, " item: ", ui$mu2RefCovariateReplaceDataFrame$covariate[i])
+
            .datEnv$i <- i
            .tmp <- try(with(.datEnv,
                         with(data,
@@ -70,7 +180,8 @@ mu2env$expit <- rxode2::expit
                        silent=TRUE)
            if (inherits(.tmp, "try-error")) {
              if (is.null(.datEnv$symengine)) {
-               .minfo(paste0("loading model to look for ", .mu3, "references"))
+               .minfo(paste0("loading model to look for ", .mu3, "/", .mu4,
+                             " references"))
                .datEnv$symengine <- ui$loadPruneSaem
                .minfo("done")
              }
@@ -81,6 +192,19 @@ mu2env$expit <- rxode2::expit
                .minfo(.txt)
                # Will put into the fit information
                warning(.txt, call.=FALSE)
+             } else {
+               .tmp <- try(.uiGetMu4(data, .datEnv,
+                                     ui$mu2RefCovariateReplaceDataFrame$covariate[i]),
+                           silent=TRUE)
+               if (!inherits(.tmp, "try-error")) {
+                 .txt <- paste0(.mu4, " item: ", ui$mu2RefCovariateReplaceDataFrame$covariate[i])
+                 .minfo(.txt)
+                 warning(.txt, call.=FALSE)
+               } else {
+                 .txt <- paste0("not ",.mu2,", ", .mu3, " or ", .mu4," item: ", ui$mu2RefCovariateReplaceDataFrame$covariate[i])
+                 .minfo(.txt)
+                 warning(.txt, call.=FALSE)
+               }
              }
            } else {
              .txt <- paste0(.mu2, " item: ", ui$mu2RefCovariateReplaceDataFrame$covariate[i])
@@ -94,7 +218,7 @@ mu2env$expit <- rxode2::expit
              .old <- str2lang(ui$mu2RefCovariateReplaceDataFrame$modelExpression[i])
              .datEnv$model <- .uiModifyForCovsRep(.datEnv$model, .old, .new)
            } else {
-             .txt <- paste0("not ",.mu2," or ", .mu3, " item: ", ui$mu2RefCovariateReplaceDataFrame$covariate[i])
+             .txt <- paste0("not ",.mu2,", ", .mu3, " or ", .mu4," item: ", ui$mu2RefCovariateReplaceDataFrame$covariate[i])
              .minfo(.txt)
              warning(.txt, call.=FALSE)
            }